backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
openpyxl==3.1.2
pandas==2.2.0
PyJWT==2.8.0
